+++ src/librand/distributions/mod.rs
@@ -20,8 +20,12 @@ that do not need to record state.
 
 */
 
-use std::num;
-use std::num::CheckedAdd;
+#![experimental]
+
+use core::prelude::*;
+use core::num;
+use core::num::CheckedAdd;
+
 use {Rng, Rand};
 
 pub use self::range::Range;
@@ -89,30 +93,32 @@ pub struct Weighted<T> {
 /// # Example
 ///
 /// ```rust
-/// use rand::distributions::{Weighted, WeightedChoice, IndependentSample};
+/// use std::rand;
+/// use std::rand::distributions::{Weighted, WeightedChoice, IndependentSample};
 ///
-/// let wc = WeightedChoice::new(vec!(Weighted { weight: 2, item: 'a' },
-///                                   Weighted { weight: 4, item: 'b' },
-///                                   Weighted { weight: 1, item: 'c' }));
+/// let mut items = vec!(Weighted { weight: 2, item: 'a' },
+///                      Weighted { weight: 4, item: 'b' },
+///                      Weighted { weight: 1, item: 'c' });
+/// let wc = WeightedChoice::new(items.as_mut_slice());
 /// let mut rng = rand::task_rng();
 /// for _ in range(0, 16) {
 ///      // on average prints 'a' 4 times, 'b' 8 and 'c' twice.
 ///      println!("{}", wc.ind_sample(&mut rng));
 /// }
 /// ```
-pub struct WeightedChoice<T> {
-    items: Vec<Weighted<T>>,
+pub struct WeightedChoice<'a, T> {
+    items: &'a mut [Weighted<T>],
     weight_range: Range<uint>
 }
 
-impl<T: Clone> WeightedChoice<T> {
+impl<'a, T: Clone> WeightedChoice<'a, T> {
     /// Create a new `WeightedChoice`.
     ///
     /// Fails if:
     /// - `v` is empty
     /// - the total weight is 0
     /// - the total weight is larger than a `uint` can contain.
-    pub fn new(mut items: Vec<Weighted<T>>) -> WeightedChoice<T> {
+    pub fn new<'a>(items: &'a mut [Weighted<T>]) -> WeightedChoice<'a, T> {
         // strictly speaking, this is subsumed by the total weight == 0 case
         assert!(!items.is_empty(), "WeightedChoice::new called with no items");
 
@@ -122,9 +128,11 @@ impl<T: Clone> WeightedChoice<T> {
         // weights so we can binary search. This *could* drop elements
         // with weight == 0 as an optimisation.
         for item in items.mut_iter() {
-            running_total = running_total.checked_add(&item.weight)
-                .expect("WeightedChoice::new called with a total weight larger \
-                        than a uint can contain");
+            running_total = match running_total.checked_add(&item.weight) {
+                Some(n) => n,
+                None => fail!("WeightedChoice::new called with a total weight \
+                               larger than a uint can contain")
+            };
 
             item.weight = running_total;
         }
@@ -139,11 +147,11 @@ impl<T: Clone> WeightedChoice<T> {
     }
 }
 
-impl<T: Clone> Sample<T> for WeightedChoice<T> {
+impl<'a, T: Clone> Sample<T> for WeightedChoice<'a, T> {
     fn sample<R: Rng>(&mut self, rng: &mut R) -> T { self.ind_sample(rng) }
 }
 
-impl<T: Clone> IndependentSample<T> for WeightedChoice<T> {
+impl<'a, T: Clone> IndependentSample<T> for WeightedChoice<'a, T> {
     fn ind_sample<R: Rng>(&self, rng: &mut R) -> T {
         // we want to find the first element that has cumulative
         // weight > sample_weight, which we do by binary since the
@@ -153,8 +161,8 @@ impl<T: Clone> IndependentSample<T> for WeightedChoice<T> {
         let sample_weight = self.weight_range.ind_sample(rng);
 
         // short circuit when it's the first item
-        if sample_weight < self.items.get(0).weight {
-            return self.items.get(0).item.clone();
+        if sample_weight < self.items[0].weight {
+            return self.items[0].item.clone();
         }
 
         let mut idx = 0;
@@ -169,7 +177,7 @@ impl<T: Clone> IndependentSample<T> for WeightedChoice<T> {
         // one is exactly the total weight.)
         while modifier > 1 {
             let i = idx + modifier / 2;
-            if self.items.get(i).weight <= sample_weight {
+            if self.items[i].weight <= sample_weight {
                 // we're small, so look to the right, but allow this
                 // exact element still.
                 idx = i;
@@ -182,7 +190,7 @@ impl<T: Clone> IndependentSample<T> for WeightedChoice<T> {
             }
             modifier /= 2;
         }
-        return self.items.get(idx + 1).item.clone();
+        return self.items[idx + 1].item.clone();
     }
 }
 
@@ -247,7 +255,9 @@ fn ziggurat<R:Rng>(
 
 #[cfg(test)]
 mod tests {
-    use {task_rng, Rng, Rand};
+    use std::prelude::*;
+
+    use {Rng, Rand};
     use super::{RandSample, WeightedChoice, Weighted, Sample, IndependentSample};
 
     #[deriving(Eq, Show)]
@@ -274,8 +284,8 @@ mod tests {
     fn test_rand_sample() {
         let mut rand_sample = RandSample::<ConstRand>;
 
-        assert_eq!(rand_sample.sample(&mut task_rng()), ConstRand(0));
-        assert_eq!(rand_sample.ind_sample(&mut task_rng()), ConstRand(0));
+        assert_eq!(rand_sample.sample(&mut ::test::rng()), ConstRand(0));
+        assert_eq!(rand_sample.ind_sample(&mut ::test::rng()), ConstRand(0));
     }
     #[test]
     fn test_weighted_choice() {
@@ -286,7 +296,8 @@ mod tests {
 
         macro_rules! t (
             ($items:expr, $expected:expr) => {{
-                let wc = WeightedChoice::new($items);
+                let mut items = $items;
+                let wc = WeightedChoice::new(items.as_mut_slice());
                 let expected = $expected;
 
                 let mut rng = CountingRng { i: 0 };
@@ -332,19 +343,19 @@ mod tests {
 
     #[test] #[should_fail]
     fn test_weighted_choice_no_items() {
-        WeightedChoice::<int>::new(vec!());
+        WeightedChoice::<int>::new([]);
     }
     #[test] #[should_fail]
     fn test_weighted_choice_zero_weight() {
-        WeightedChoice::new(vec!(Weighted { weight: 0, item: 0},
-                                 Weighted { weight: 0, item: 1}));
+        WeightedChoice::new(&mut [Weighted { weight: 0, item: 0},
+                                  Weighted { weight: 0, item: 1}]);
     }
     #[test] #[should_fail]
     fn test_weighted_choice_weight_overflows() {
         let x = (-1) as uint / 2; // x + x + 2 is the overflow
-        WeightedChoice::new(vec!(Weighted { weight: x, item: 0 },
-                                 Weighted { weight: 1, item: 1 },
-                                 Weighted { weight: x, item: 2 },
-                                 Weighted { weight: 1, item: 3 }));
+        WeightedChoice::new(&mut [Weighted { weight: x, item: 0 },
+                                  Weighted { weight: 1, item: 1 },
+                                  Weighted { weight: x, item: 2 },
+                                  Weighted { weight: 1, item: 3 }]);
     }
 }
