+++ src/libserialize/base64.rs
@@ -261,7 +261,6 @@ impl<'a> FromBase64 for &'a str {
 #[cfg(test)]
 mod tests {
     extern crate test;
-    extern crate rand;
     use self::test::Bencher;
     use base64::{Config, FromBase64, ToBase64, STANDARD, URL_SAFE};
 
@@ -335,7 +334,7 @@ mod tests {
 
     #[test]
     fn test_base64_random() {
-        use self::rand::{task_rng, random, Rng};
+        use std::rand::{task_rng, random, Rng};
 
         for _ in range(0, 1000) {
             let times = task_rng().gen_range(1u, 100);
