+++ src/test/compile-fail/task-rng-isnt-sendable.rs
@@ -10,11 +10,11 @@
 
 // ensure that the TaskRng isn't/doesn't become accidentally sendable.
 
-extern crate rand;
+use std::rand;
 
 fn test_send<S: Send>() {}
 
 pub fn main() {
-    test_send::<::rand::TaskRng>();
-    //~^ ERROR: incompatible type `rand::TaskRng`, which does not fulfill `Send`
+    test_send::<rand::TaskRng>();
+    //~^ ERROR: incompatible type `std::rand::TaskRng`, which does not fulfill `Send`
 }
